Fix issue #178
To fix issue #156, we were exluding all directories, so that only the swift files get added.
But the fix for that issue, lead to a bug where xcdatamodeld and xcassets folder
references were not being added to the pbxproj.
Adding a check to not reject these two directories for include paths fixes this bug.